refactor to use a single settings.py, allow a settings.yaml
