+++ shub/plugins/google_build/__init__.py
@@ -5,8 +5,8 @@ from shub.logger import bot
 # Ensure that application credentials exist
 for required in [
     "GOOGLE_APPLICATION_CREDENTIALS",
-    "SREGISTRY_GOOGLE_PROJECT",
-    "SREGISTRY_GOOGLE_BUILD_EXPIRE_SECONDS",
+    "GOOGLE_PROJECT",
+    "GOOGLE_BUILD_EXPIRE_SECONDS",
 ]:
     if not hasattr(settings, required):
         bot.exit("%s not defined in secrets." % required)
