+++ shub/plugins/google_build/github.py
@@ -24,7 +24,7 @@ from shub.settings import (
     DISABLE_BUILDING,
     DISABLE_GITHUB,
     DOMAIN_NAME,
-    SREGISTRY_GOOGLE_BUILD_LIMIT,
+    GOOGLE_BUILD_LIMIT,
 )
 
 from .utils import (
@@ -467,7 +467,7 @@ def verify_payload(request, collection):
     if is_over_limit():
         message = (
             "Registry concurrent build limit is "
-            + "%s" % SREGISTRY_GOOGLE_BUILD_LIMIT
+            + "%s" % GOOGLE_BUILD_LIMIT
             + ". Please try again later."
         )
