+++ shub/plugins/google_build/views.py
@@ -31,7 +31,7 @@ from shub.settings import (
     DISABLE_BUILD_RECEIVE,
     DISABLE_BUILDING,
     DISABLE_GITHUB,
-    SREGISTRY_GOOGLE_BUILD_LIMIT,
+    GOOGLE_BUILD_LIMIT,
 )
 from shub.settings import VIEW_RATE_LIMIT as rl_rate
 from shub.settings import VIEW_RATE_LIMIT_BLOCK as rl_block
@@ -211,7 +211,7 @@ class RecipePushViewSet(ModelViewSet):
         if is_over_limit():
             message = (
                 "Registry concurrent build limit is "
-                + "%s" % SREGISTRY_GOOGLE_BUILD_LIMIT
+                + "%s" % GOOGLE_BUILD_LIMIT
                 + ". Please try again later."
             )
             print(message)
