--- shub/settings/api.py
@@ -1,32 +1,0 @@
-"""
-
-Copyright (C) 2017-2022 Vanessa Sochat.
-
-This Source Code Form is subject to the terms of the
-Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
-with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-"""
-
-
-REST_FRAMEWORK = {
-    # You can require authentication for your API
-    #'DEFAULT_PERMISSION_CLASSES': (
-    #    'rest_framework.permissions.IsAuthenticated',
-    # ),
-    "DEFAULT_SCHEMA_CLASS": "rest_framework.schemas.coreapi.AutoSchema",
-    "DEFAULT_PAGINATION_CLASS": "rest_framework.pagination.LimitOffsetPagination",
-    "DEFAULT_AUTHENTICATION_CLASSES": (
-        "rest_framework.authentication.SessionAuthentication",
-        "rest_framework.authentication.TokenAuthentication",
-        "rest_framework.authentication.BasicAuthentication",
-    ),
-    # You can also customize the throttle rates, for anon and users
-    "DEFAULT_THROTTLE_CLASSES": ("rest_framework.throttling.AnonRateThrottle",),
-    # https://www.django-rest-framework.org/api-guide/throttling/
-    "DEFAULT_THROTTLE_RATES": {"anon": "100/day", "user": "1000/day"},
-    "PAGE_SIZE": 10,
-}
-
-
-API_VERSION = "v1"
