--- shub/settings/logging.py
@@ -1,25 +1,0 @@
-"""
-
-Copyright (C) 2017-2022 Vanessa Sochat.
-
-This Source Code Form is subject to the terms of the
-Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
-with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-"""
-
-import os
-
-# Default Django logging is WARNINGS+ to console
-# so visible via docker-compose logs uwsgi
-LOGGING = {
-    "version": 1,
-    "disable_existing_loggers": False,
-    "handlers": {"console": {"class": "logging.StreamHandler"}},
-    "loggers": {
-        "django": {
-            "handlers": ["console"],
-            "level": os.getenv("DJANGO_LOG_LEVEL", "WARNING"),
-        }
-    },
-}
